pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shorts"
version = "0.1.0"
description = "Chinese short video to English YouTube automation pipeline"
readme = "PLAN.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.54.0",
  "pysubs2>=1.7.3",
  "PyYAML>=6.0.2",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
asr = [
  "faster-whisper>=1.1.1",
]
dev = [
  "ruff>=0.13.0",
]

[tool.setuptools]
packages = ["src", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py311"

